requirements_hpc.txt
Click==7.0
comet-ml==1.0.55
h5py==2.10.0
joblib==0.14
Markdown==3.1
mock==3.0.5
numpy==1.16.3
pandas
pytest==4.6.1
python-dateutil==2.8.0
scikit-learn==0.22
scipy
tabulate==0.8.3
tqdm==4.32.1
tensorflow-addons==0.14.0



